Added Script For Changing Settings
Added a settings button script that the buttons will call to set the game settings. This is deone because of the static game settings object not being able to be pointed to by the buttons upon loading the scene. The settings button script finds the game settings script and passes along the settings.