Started UI Designs
Started to add art and styles to the UI. Need to come up with a good background picture. Also need to settle on an asthetic. Real or Cartoony? Also fixed a problem where I broke the settings. The buttons need to use their parent's settings button object so the script can find the static game settings object.