Added More Verses. Fixed UI Scaling and High Score Bug
Added more Verses to Easy, Hard, and Medium. Put them in the game and tested. Fixed a problem with images in the UI not scaling properly and looking blurry. Fixed an issue with the high score not updating after a game because playerprefs weren't saving.  Improved text spacing for verses on the index card.